refactor: break plugin pod spec builder
