+++ pkg/plugins/trivy/image_test.go
@@ -1,0 +1,49 @@
+package trivy_test
+
+import (
+	"testing"
+
+	"github.com/aquasecurity/trivy-operator/pkg/plugins/trivy"
+	"github.com/stretchr/testify/assert"
+	"github.com/stretchr/testify/require"
+)
+
+func TestGetMirroredImage(t *testing.T) {
+	testCases := []struct {
+		name          string
+		image         string
+		mirrors       map[string]string
+		expected      string
+		expectedError string
+	}{
+		{
+			name:     "Mirror not match",
+			image:    "alpine",
+			mirrors:  map[string]string{"gcr.io": "mirror.io"},
+			expected: "alpine",
+		},
+		{
+			name:     "Mirror match",
+			image:    "alpine",
+			mirrors:  map[string]string{"index.docker.io": "mirror.io"},
+			expected: "mirror.io/library/alpine:latest",
+		},
+		{
+			name:          "Broken image",
+			image:         "alpine@sha256:broken",
+			expectedError: "could not parse reference: alpine@sha256:broken",
+		},
+	}
+
+	for _, tc := range testCases {
+		t.Run(tc.name, func(t *testing.T) {
+			expected, err := trivy.GetMirroredImage(tc.image, tc.mirrors)
+			if tc.expectedError != "" {
+				require.EqualError(t, err, tc.expectedError)
+			} else {
+				require.NoError(t, err)
+				assert.Equal(t, tc.expected, expected)
+			}
+		})
+	}
+}
